+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 
 if (process.env.NEXT_RUNTIME === "nodejs" && process.env.NODE_ENV === "development") {
-  const { worker } = await import("../../mocks/server");
+  const { worker } = await import("../mocks/server");
 
   worker.listen();
 }
